rpk topic describe: only print the stable column if requested with --stable
The stable column is different from the HWM in two cases:
1, and far more common: Records were produced between the HWM and LSO.
2, less common: Transactions are being used.

We always requested the stable offset after HWM, so frequently, the
stable offset would look higher than the HWM (which is wrong!), and it
would actually be equal behind the scenes but only look different due to
time. Hiding it by default is less confusing, and requesting the stable
offsets second rather than third prevents stable appearing higher.