archival: Disable retention calc if archive is present
The log is split into two parts, archive and STM. The
retention_calculator computes retention parameters for the STM part of
the log. It shouldn't be used in case if there is an archive part
present. In this case retention on the archive should be completed
first. Otherwise we might introduce a gap between the archive and STM.
This commit adds a safeguard to prevent this.