archival: Do retention and GC using 'async_manifest_view'
Update 'ntp_archiver' to use the 'async_manifest_view'. The housekeeping
loop can now do retention in the archive or in the STM area of the log.
The first one is implemented using 'async_manifest_view' that can
compute retention parameters and also find candidates for removal by GC.

cloud_storage: Use 'async_manifest_view' in the read path

The component is used in the 'remote_partition' The 'remote_segment'
had to be changed to avoid passing the manifest to the c-tor.
The unit tests are updated to initialize both 'remote_partition' and
'remote_segment' correctly. Also, the 'cloud_storage_e2e_test' is
updated to actually read from the archive section of the log.