add feature to trigger managed rolling decommission
This adds the ability to do planned decommissioning of StatefulSet pods.

To trigger a managed decommission, set an annotation on the Cluster resource with an RFC3339 deadline.

```
kubectl annotate cluster managed-decommission operator.redpanda.com/managed-decommission="2999-12-31T00:00:00Z"
```

The Cluster controller will decommission each pod from 0 to N, delete the pvcs and the pod, waiting for the replacement pod to join and the cluster to be healthy between each deletion.

When finished, the controller will remove the annotation.

Should the deadline pass during the managed decommission, the controller will not continue decommissioning pods. To resume the decommission, set the annotation to a timestamp in the future

(cherry picked from commit 5e44461ed94a9687145aebfe0435cf17a756a0a0)