cloud_storage: Update the partition_manifest
Add the list of spillover manifests to the partition manifest. The list
is supposed to be used instead of the ListObjectsV2 api in S3. The
'segment_meta' structure is used to represent individual manifests.
The compressed column-store which is used to store segments is also used
to store spillover manifests.