c/partition_balancer: add partition counts rebalancing operator
The algorithm is simple: just try to move every replica. Assuming that
due to the min-count soft constraint it will end on a node with fewer
replicas than the original one, we will move towards (local) optimum.

This operator is triggered only if we have non-empty
"nodes-to-rebalance" collection in partition_balancer_state. It will be
updated by members_manager and will contain all nodes that have been
added but haven't yet been issued a corresponoding finish_reallocations
command.