storage: move ownership of disk paths to storage::node
Initializes storage::node with paths to the data/cache disks, and forces
access to these disk to be specified in terms of the type of disk rather
than a bare string path. The bare string paths are difficult to work
with when locations like admin server want to interact with the stat
information about disks, and don't care about the path.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>