archival: fix collection for adjacent reupload
Consider the log configuration below. Assume that the adjacent segment
reupload has decided to merge cloud segments [104, 113] and [113, 115].
The `segment_collector` should produce an upload candidate that covers
offets [104, 115]. Due to a bug fixed in this PR, the upload candidate
produced would span the [2, 115] offset range. This caused the merge to
fail.

The problems stems from the fact that compacted segment reupload can be
more lax about the offset range provided to the `segment_collector`. The
reupload can cover offsets below the provided `start_offset` if that's
in the middle of a segment. However, that's not the case for calls into
`segment_collector` made by the segment merger. The upload candidate
needs to match the range of the merge precisely.

This commit tweaks disables the offset fudging in `segment_collector`
when the specified mode is
`segment_collector_mode::collect_non_compacted`.
        +-----------------------------------++---------------------+
Local   |2                               115||116               211|
        +-----------------------------------++---------------------+
        +----------++----------++-----------++---------------------+
Cloud   |2      103||104    113||114     115||116               211|
        +----------++----------++-----------++---------------------+

Fixes #10583

Minor non-functional changes were made to the unit test in order for it
to build.

(cherry picked from commit 23971b9ef739c90f78c193795b358e9260f24933)