rpk: list virtual cluster on cloud login
Now rpk cloud login will display a list of
virtual clusters and clusters when you succesfully
logged in. Then you can pick one of them and rpk
will create a profile with the selection.

If there is just one cluster, it will select that
cluster by default and create a profile.