e2e_shadow_indexing_test: bump workload
The test could sometimes fail to upload a sufficient number of segments
(e.g. 17k instead of ~25k). It seems to be caused by the fact that
segment rolling will not always keep up with the produce rate, resulting
in fewer, larger segments than would be calculated by the data and
segment size.

This commit bumps the total data size written by an order of magnitude
-- the workload won't finish given the test stops the worklaod once a
sufficient number of segments is reached.