storage: take _concurrent_flushes units before _pending_head_write units
In all cases of taking _concurrent_flushes and accessing
_pending_head_write (even just checking available units, rather than
acquiring the units) together, _concurrent_flushes is taken first,
except when we fallocate.

This seems like a flavor of lock order inversion that can be solved by
consistent ordering of locks. So, this patch tweaks the ordering to be
consistent with other accesses: first taken _concurrent_flushes, then
_pending_head_write.