pandaproxy: Support SCRAM-SHA-512 for basic_auth
User credentials provided by basic auth cannot provide a sasl_mechanism.

During request authorization, both mechanisms are tried, but the
mechanism that succeeds was not propagated in the request_auth_result.

This commit propagates the successful mechanism and uses that during
the creation of the client when http basic auth is used for requests
to Pandaproxy, instead of hardcoding it to SCRAM-SHA-256.

Signed-off-by: Ben Pope <ben@redpanda.com>