ducktape: Change condition for upload check
The condition for the test checks if the number of segments uploaded is
exactly equal to one minus the local segment count. Since the max
upload interval is set we can end up uploading all segments to the
cloud (as seen in CI failures). This change adjusts the success
condition to include the case where all segments are uploaded.