cloud_storage: tweak test to use default archiver
We have been noticing failures of test_produce_consume_from_cloud while
waiting for the intial segment to be uploaded. Looking at the logs, tuns
out that the inlined call into the archiver was racing with the usual
upload loop. The wait condition has been changed to check that the
manifest contains exactly one segment.

Fixes #11091