ducktape: Adjust NTP size calculation below start
The cloud log size for NTP helper adds up the sizes of meta objects from
a manifest. This method is changed to accept a parameter to ignore those
segments which are below the manifest start offset, in order to provide
a more accurate size in cases where comparisons are performed with on
disk sizes.