tests: fixed setting up consumer for interruption test
Previously the producer used high throughput which may lead to the
situation in which consumer continued without a gap in the offset space
even after the topic was deleted.
The producer might have already produced enough records to fill the gap
and allow consumer to continue where it left of.

Using lower producer throughput and waiting for large number of records
to be consumed before deleting topic will prevent this situation from
happening.

Fixes: #11459

Signed-off-by: Michal Maslanka <michal@redpanda.com>