cloud_storage: be less restrictive in moving start kafka offset
DeleteRecords may be called on an offset that has yet to be uploaded,
and the cloud storage layer must remember this. To that end, this commit
reduces the restrictions we previously had on the offset: it is now
legal for the offset to be outside the cloud storage bounds. It is now a
simple monotonically increasing offset.

The goal here is to decouple any overarching partition-wide desired
start offset from what is actually in cloud storage.

Semantically, the _start_kafka_offset changes moving forward: cloud
storage will no longer automatically consider the offset when seeking
around or reporting the start offset, instead relying on callers to do
this (coming in a following commit).