memory-sampling: Disable top-n alloc sites logging on low memory
We added a condition variable in the batch cache which would notify the
memory sampling service whenever a reclaim happened.

This is bad because `condition_variable.signal()` can potentially
allocate which can lead to recursive reclaim calls.

Disable the notification (and as a consequence the logging when reaching
a lower watermark threshold) until we have a proper fix. We will either
need to use another notification mechanism or notify from the background
reclaimer.

Keeping all the other code in place as we will want to reuse it.