kafka: clamp the start offset in Kafka layer
Previous commits removed explicit clamping of Kafka start offset in
storage and cloud storage. This commit updates the clamping back at the
Kafka layer.

Each storage layers' start offset methods are used for more than just
reporting, e.g. deciding whether a given offset falls in local or cloud
storage, or in the STM or spillover cloud region. The intent here is to
reduce the complexity of the storage/cloud layers having to account for
the override in its reporting of offsets.

To do this, this exposes pieces of the log_eviction_stm to be composed
at the Kafka layer.