memory-sampling: Move low watermark logging to be timer based
This patch moves the available memory low watermark checking from being
notification based to being checked on a timer.

We had originally discarded this idea because it's fundamentally not
great. If we were to alloc really fast then we would possibly miss
intermediate watermarks or all watermarks altogether before OOMing.

Rather we'd want something notification based that notifies us when the
individual watermarks are actually reached.

seastar doesn't offer anything like this. Instead we made the batch
cache be the notifier. The batch cache is closely linked to the
allocator as it registers a reclaim hook which the allocator will call
when under memory pressure. Hence, the batch cache sounds like a good
place to put this.

However, there is actually multiple problems with this:

 - Allocating from the reclaim hook: We absolutely want to avoid
   allocating from the reclaim hook as otherwise we might recurse into
   the reclaim hook and this is brittle in general. Note it's very easy to
   allocate so extra care must be taken.
 - Reclaim doesn't change available memory: Available memory is defined
   as the sum of seastar free memory plus what is reclaimable by the
   batch cache. Hence, on reclaim memory just moves from reclaimable to
   free. It's still good as a notification mechanism. Reclaim only gets
   invoked on memory pressure so probably the next time reclaim gets
   invoked available memory is now lower as an allocation happened as a
   follow up to the previous reclaim.
 - Depending on the size of the batch cache you might not get a
   notification at higher levels of available memory (e.g.: 50%
   available) as there is no memory pressure in the allocator. This again
   is probably fine in practice but again just opens it up for edge cases.
 - Background reclaim: The batch cache has an additional form of reclaim
   which is driven by the background reclaimer. This gets triggered
   whenever we put something into the batch cache. Again this is probably
   not the right place, putting something into the batch cache makes
   available memory go up but we are interested in a change in the opposite
   direction.

One thing to keep in mind is that the lower watermark logging is the
least important part of the memory sampling implementation. The more
important ones are on-OOM logging and on-demand retrieval both of which
are unaffected by any of this.

Hence, in the spirit of keeping things simple and not further
complicating the brittle batch cache reclaim mechanism we move this back
to be timer based. With a 60s check frequency we will still catch any
form of slow over time memory leak.

Note this changeset looks rather huge from a line count but most of that
is just undoing changes to all the tests that had to be touched because
of having to pass down the mem sampling service.