archival: fix infinite loop in archival gc
In order to queue segments up for deletion we iterate over the spillover
manifests and pick the ones below the archive start offset. This
procedure could result in an infinite loop if the archive start offset
equals the start offset of the STM region. Checking if the cursor has
reached the end solves the issue.