cloud_storage: Clean up segment hydration
When segment is hydrated, we are guaranteed to be in legacy mode, which
means that we were not able to download an index from cloud storage.
The change cleans up the code path in segment hydration, so that if a
segment is downloaded, an index is always created on the fly from the
segment data.