c/rm_stm: do not persist evicted producer ids in snapshot
Only store those producer id sequences which offset is
greater than log start offset. This way a snapshot will
not retain producers ids for which all the batches were
removed with log cleanup policy.

Note that we are not removing producer ids from the in
memory state but rather relay on the expiration policy
to do it, however when recovering state from the
snapshot removed producers will be gone.

Signed-off-by: Michal Maslanka <michal@redpanda.com>