cloud_storage: account for start override in STM manifest retention
We previously accounted for the start override by incorporating it into
the start offset at the time of applying update_start_kafka_offset
batch. Now that this has changed, we need to apply retention on the STM
region of the manifest.

This commit does so by introducing a new retention strategy to the
retention calculator class.