raft/recovery_throttle: use total available capacity.
When computing a shard's target rate based on it's share of deficit,
we are losing some rate while rounding the deficit ratio to a size_t.

Example: A rate of 10 when shared between two shards with 25% and 75%
deficit shares was resulting in a total rate assignment of

floor(25/100 * 10) + floor(75/100 * 10) = 9

effectively losing 1 unit of rate.

This patch accumulates all such lost capacity and assigns to the last shard.
This is not a problem in practice but tests use this total rate check
(sum of per shard rates) to make sure the new rate has been applied.
This patch ensures that invariant is maintained.

(cherry picked from commit b703296a3323fd7c822ccfcb2eeda7f0211ecfb6)