kafka: bump timequery results up to start override
The cluster::partition abstraction is mostly unaware of Kafka start
overrides when it comes to fetching data. As such, a timequery may
return a result that falls below the start override.

To avoid returning an unqueryable offset, this patch updates the
behavior to check the timequery result against the override, and
performs a read at the override offset if the timequery result was too
low.