storage: add control loop to monitor log storage usage
The purpose of this configuration option is to be analgous to the cloud
storage cache's max size option, namely, that used space should stay
below this threshold.

When a value is not set the behavior is unchanged and there is no limit.

Computes and broadcasts a flag that indicates that max size has been
exceeded.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>