storage: on max size exceeded run gc immediately
This threads a reference to the high-level api into storage allowing the
houskeeping loop to inspect the state of max bytes being exceeded. When
it is exceeded housekeeping will priortizie GC. The storage space
monitor will also pokethe housekeeping thread.

Threading the api down into storage is also going to be used in later
work that allows logs to preemptively wake up the monitor when it
appears that we are getting close to the max bytes threshold.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>