storage: add storage max size high watermark
Because we want to run close to the maximum size, we need to avoid the
situation where the system is blocking/unblocking produces frequently.
To deal with this we introduce a high watermark which defines a
configurable threshold that is less than the maximum size, which acts as
the trigger for starting the gc process and any other mitigation to
avoid reaching the maximum size.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>