storage: reduce memory footprint of compaction_backlog
It was unnecessary to accumulate all segments in the topic,
since each term is calculated separately.  Do the per-term
calculation inline.

This still has a memory footprint that scales linearly with
the number of segments in a term, but not the total number
of segments in a partition.

(cherry picked from commit 73619ebe7af3fed23717560ea98e4cc27ec2d1fc)