tests: fix quiesce call in topic deletion test
This fixes commit 8355e4ec4593c852401c3ade97871da52d5e04b7

quiesce_uploads would fail silently if you called it for
a topic that didn't exist: that's fixed in the previous
commit.

This commit fixes the underlying issue, that we were
passing a string instead of a list of strings, so
quiesce_upload was trying to wait for each character
in the string as if it was a topic name.

Fixes https://github.com/redpanda-data/redpanda/issues/8496
Fixes https://github.com/redpanda-data/redpanda/issues/10655
Fixes https://github.com/redpanda-data/redpanda/issues/9629