cloud_storage_e2e_test: begin using new test methods
Some tests do not expect to use the upload loop, instead relying on the
test body to drive the archiver. This commit transitions one such test
to using test-only configs and methods, which should make the test more
robust.

Along the way, this updates the workload to use the new
remote_segment_generator.