archival/scrubber: retry scrub deletion requests
Previously, the scrubber did not re-send retryable requests. The
intention here was to drop if the cloud storage provider is applying
throttling as it is unlikely that much progress would be made.

However, the connection is killed after specific request sequences
(posibly due to htpp client bug). Not retrying means that the scrubber
does not make progress. This patch changes the retry strategy used for
purging partitions to exponential backoff in order to work around the
issue.