cloud_storage: Read offset arrays when creating coarse index
When creating coarse index, in addition to reading the index fields, we
also read the offset fields. This ensures that in cases where the index
is very small and all the data is contained in the offset fields, a
mini index can still be generated.

Additionally the calculation of offsets to add in the index is changed
so that we do not end up skipping offsets when the mod of current
offset is unchanged, which is a bug that the previous iteration of the
code was susceptible to.