tests: fix racy behavior in `drop_lifecycle_marker_test`
Because a scrubber iteration was likely to be in flight
at the time we drop the lifecycle marker, there is a good
chance that Redpanda will end up _partially_ deleting
the topic when we unblock the firewall at the end of
the test, causing failures on the end-of-test validation
of cloud storage consistency.

Work around this by restarting redpanda before unblocking
the firewall: this effectively acts as a barrier in the scrubber
loop.

Fixes https://github.com/redpanda-data/redpanda/issues/11232