r/consensus: do not update heartbeat timestamp with prevote request
When a voter receives vote request and it votes for the candidate it
updates the last heartbeat timeout. If this happens during the prevote
phase and in a deployment with even number of locks it may lead to
temporary live lock and not being able to elect the leader.

Fixes: #11657

Signed-off-by: Michal Maslanka <michal@redpanda.com>
(cherry picked from commit 995c7750fdcd6315785c1b982dc415979facb31e)