storage: add log interface for external space management
Space management control requires a few things:

1. Access segment information for each partition that allows it to
   reason about how much of a partition should (and can) be truncated to
   meet a target reclaim size.

2. A mechanism for overriding local retention rules for cloud enabled
   topics. Data that is uploaded into the cloud, independent of a topics
   retention settings, can be removed safely on disk. But choose a
   policy wisely.

This commit adds (1) by exposing raw segments from the log, so the
caller should be careful. The flip side is that its just expected to
look at size information and offsets.

For (2) the normal GC path is hijacked when a specific offset is
requested. In this case the normal size/time based retention is ignored
and the requested offset is used. The requested offset is cleared after
GC runs. This makes sense because after the data has been reclaimed
reads may rehydrate storage at which point space management may
reconsider the data.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>