rptest: make kgo seq consumer read full log
Previously, whenever a KgoVerifierSequentialConsumer and a
KgoVerifierProducer ran in parallel (which a number of tests do), there
was no guarantee that the consumer would read all partitions entirely.
Whenever `wait` is called on the consumer, the `last_pass` command is
triggered without taking into account if the producer has completed.

This is problematic because it can hide stuck consumer issues. This
patch makes use of the enhanced kgo-verifier status which now contains
the high watermark for each partition. When creating a sequential
consumer, a reference to the producer can be provided. If that was the
case, when `wait` is called on the consumer we will wait until the
producer has finished and the high watermarks match.