archival: use full manifest replacement in read replicas
This replaces attempting to incrementally update the
state machine based on the delta between local manifest
and remote manifest.  The incremental updates were hard
to test, and needed updating each time we changed
how the manifest worked.

This way we are doing exactly what we intend in the
simplest way: make the local state equal to the
remote state.

This makes read replicas work with metadata spilling.