archival: remove archive once fully truncated and cleaned
It doesn't appear that housekeeping can currently reset the archive
start offset to model::offset{}. This is important because
model::offset{} is used to indicate the absence of an archive (e.g. see
ntp_archiver::stm_retention_needed()).

This adds a check to partition_manifest::set_archive_clean_offset() that
will reset the manifest upon being truncated to the start of the STM
manifest.