ducktape: Adjust small cache test for throttling
The test which simulates a cache smaller than the segment size may
suffer from timeout failures due to throttling of cache trim. To address
this the chunk size is made smaller to reduce trimming frequency, and we
read fewer messages in the consumer, just enough that the chunk read
path is exercised.