storage: clamp raft retention to 1hr for read replicas
A read replica topic will be created with the default
7 hour retention, but because the partitions in such a
topic are only used for configuration batches, it is
not useful to retain any data beyond the most recent
snapshot.

Logically our retention time would be zero, but I don't
want the GC code to constantly try to trim, so use
1 hour for a balance between not retaining much log,
and not trying to GC gratuitously frequently.

Also set segment.ms to this 1 hour retention, so that
we will roll segments often enough to promptly
do GC: these segments are very unlikely to hit
size thresholds to prompt a roll.