add cdt script to install dt deps on cdt infra
install-deps is currently part of the vtools repo.
CDT is designed to always run tip of vtools dev.
This causes issues when for example a dt dependency
exists in redpanda `dev` branch, and not in `v22.3.x`
branch. Once we teach install-deps of vtools/dev
to install a ducktape dependency, then it'll expect the
same dependency file in all cdt runs, including older
branches.

To avoid that, the install-deps should follow the redpanda
ducktape dependencies per branch. Additionally, developers
could modify/add a ducktape dependency and update cdt
at once

ref #redpanda-data/devprod#741