config: enable cloud_storage_spillover_manifest_size by default
Set 64kiB, this is pretty small but enough to handle a typical modest
partition's metadata with hourly uploads for a week.

The practical result will be that spilling will kick in for two
kinds of partition:
- Partitions with much longer than default retention
- High throughput partitions that write lots of segments
  because they have lots of data.

This threshold is unlikely to be met by most of our automated tests,
so it is still necessary to have explicit testing that
triggers spillover by setting smaller than default thresholds.