cloud_storage: Record client acquisition and download latencies
The measurements returned by the probe are unique ptrs. Usually at the
end of scope destruction of the unique ptr would trigger recording the
metrics into the histogram. However in this case we want to record the
metric right after the operation (client acquired or segment downloaded
or failed) has completed. So we resort to manually destroying the
measurement, triggering the recording into the histogram.