cloud_storage: fix size based retention accounting
This commit fixes a number of issues with the process of finding the
archive truncation point for size based retention.

Cosider the example below:
A      B C      D E      F G      H
[------] [------] [------] [------] ....
   10B      10B      10B   |  10B
                           |
                           |
---------------------------|-------
       Archive                STM

There's three segments in the archive and multiple in the STM region.

1. Assume we have to remove 20B to satisfy the retention policy. The
new start offset of the archive should be E. Previously, we'd pick C as
the truncation point. This is due to the fact that we were only updating
the truncation point when the segment was collected. The fix is
to update the truncation point each time (until breaking).

2. Now assume we have to remove 35B to satisfy the retention policy.
The trunaction point should be G (the start of the STM manifest).
Previously, we'd pick E instead because `eof` was not true after the
last spill manifest. The fix is to specify the end offset of the
archive, (G-1) in this case, such that the cursor returns an EOF when
trying to step in the STM region. If the cursor returns EOF, then the
entire archive was consumed and the start offset needs to be advanced to
the STM start offset.

Note: I've also updated the time based retention to work in the same
style with regards to EOF.