rptest: fix and tighten up archive retention test
Previously, the test did not perform exact checks for the number of
segments deleted by retention. A number of test issues bubbled up when
I did this, but the most significant was that the time based test wasn't
working.

This was due to the fact that segment timestamps were too bunched up and
the precise check couldn't work. To fix that, custom timestamps with a
2 minute delta between records are used.