archival: Validate archival_metadata_stm updates
Previously we saw several cases in which:
- metadata didn't match the actual uploaded segment;
- metadata was inconsistent next segment didn't connect with the
  previous one;

This commit solves this by introducing the validation mechanism to the
ntp_archiver and archival_metadata_stm.

During the segment upload ntp_archiver calculates the segment stats
(base offset, last offset, size, offset translation parameters). This
information is considered a 'ground truth' about the upload. It gets
propagated up the call chain until the 'ntp_archiver' finished uploading
the data and produced the 'segment_meta' instance that should be added
to the STM. The 'segment_meta' is validated against the previous log and
the 'ground truth' that we obtained by analyzing the segment.

The 'archival_metadata_stm' also validates every update using more
simple procedure. It checks that the new pice of metadata connects with
the previous one and that it can't replace existing segment in an unsafe
mannter.