cloud_storage: add timestamps to remote segment index
This is a prerequisite to make timequeries seek
to the proper chunk, instead of reading from the
start of segments.

The timestamp we care about is the batch max timestamp,
because timequeries will want to find a batch that is
definitely before the target timestamp, to start their
scan from there.

Related: https://github.com/redpanda-data/redpanda/issues/11801