storage: don't force must-have reservation in reporting
Storage had been protecting the latest two segments from low-disk
reclaim by not reporting them as eligible. In reality it is not the case
that we want to have this protection for every partition, but rather the
must-have guarnatee is more about aggregate level space and policies
should be free to be more aggressive when needed.

This commit still protects the active segment. It remains to be seen if
we want to report this one as well, but it is still available by having
the space management code force a roll if required (or wait for
segment.ms to expire).

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>