model/record: `futurize_invoke` `f` in `for_each_record_async`
The function passed to `for_each_record_async` might not return a future.

This is the case when for_each_record_async` is used primary for the
purpose of yeilding, rather than to accept an async function.

`futurize_invoke` the passed function so that the call site doesn't have
to be modified.

Signed-off-by: Ben Pope <ben@redpanda.com>
(cherry picked from commit e24eb396cea4ca8fec6d43f50569b2bfe34ee006)