k/replicated_partition: fixed returning `log_end_offset` for empty logs
In Kafka the semantics of `log_end_offset` is defined as a next offset
assigned to a record produced to a given partition.

When local log is empty its `dirty_offset` is equal to `start_offset -
1`. In this case a `log_end_offset` should return Kafka offset
corresponding to the `start_offset` as this is the next offset assigned
to a record produced to given topic partition.

Signed-off-by: Michal Maslanka <michal@redpanda.com>